requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
altair==4.1.0
    # via streamlit
argon2-cffi==20.1.0
    # via notebook
astor==0.8.1
    # via streamlit
async-generator==1.10
    # via nbclient
attrs==21.2.0
    # via
    #   jsonschema
    #   markdown-it-py
    #   streamlit
backcall==0.2.0
    # via ipython
base58==2.1.0
    # via streamlit
bleach==3.3.0
    # via nbconvert
blinker==1.4
    # via streamlit
cachetools==4.2.2
    # via streamlit
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via argon2-cffi
chardet==4.0.0
    # via requests
click==7.1.2
    # via streamlit
cycler==0.10.0
    # via matplotlib
debugpy==1.3.0
    # via ipykernel
decorator==5.0.9
    # via
    #   ipython
    #   validators
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.3
    # via
    #   altair
    #   nbconvert
gitdb==4.0.7
    # via gitpython
gitpython==3.1.18
    # via streamlit
idna==2.10
    # via requests
ipykernel==6.0.1
    # via
    #   ipywidgets
    #   notebook
    #   pydeck
ipython==7.25.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   notebook
    #   traitlets
ipywidgets==7.6.3
    # via pydeck
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via
    #   altair
    #   nbconvert
    #   notebook
    #   pydeck
jsonschema==3.2.0
    # via
    #   altair
    #   nbformat
jupyter-client==6.1.12
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
jupytext==1.11.3
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
markdown-it-py[plugins]==1.1.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.2
    # via seaborn
matplotlib-inline==0.1.2
    # via ipython
mdit-py-plugins==0.2.8
    # via markdown-it-py
mistune==0.8.4
    # via nbconvert
nbclient==0.5.3
    # via nbconvert
nbconvert==6.1.0
    # via notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   jupytext
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via nbclient
notebook==6.4.0
    # via
    #   -r requirements.in
    #   widgetsnbextension
numpy==1.21.0
    # via
    #   -r requirements.in
    #   altair
    #   matplotlib
    #   pandas
    #   pydeck
    #   scipy
    #   seaborn
    #   streamlit
packaging==21.0
    # via
    #   bleach
    #   streamlit
pandas==1.3.0
    # via
    #   -r requirements.in
    #   altair
    #   seaborn
    #   streamlit
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via
    #   matplotlib
    #   streamlit
pipdeptree==2.0.0
    # via -r requirements.in
prometheus-client==0.11.0
    # via notebook
prompt-toolkit==3.0.19
    # via ipython
protobuf==3.17.3
    # via streamlit
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pycparser==2.20
    # via cffi
pydeck==0.6.2
    # via streamlit
pygments==2.9.0
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   streamlit
pytz==2021.1
    # via
    #   pandas
    #   tzlocal
pyyaml==5.4.1
    # via jupytext
pyzmq==22.1.0
    # via
    #   jupyter-client
    #   notebook
requests==2.25.1
    # via streamlit
scipy==1.7.0
    # via seaborn
seaborn==0.11.1
    # via -r requirements.in
send2trash==1.7.1
    # via notebook
six==1.16.0
    # via
    #   argon2-cffi
    #   bleach
    #   cycler
    #   jsonschema
    #   protobuf
    #   python-dateutil
    #   validators
smmap==4.0.0
    # via gitdb
streamlit==0.84.0
    # via -r requirements.in
terminado==0.10.1
    # via notebook
testpath==0.5.0
    # via nbconvert
toml==0.10.2
    # via
    #   jupytext
    #   streamlit
toolz==0.11.1
    # via altair
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   streamlit
    #   terminado
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   pydeck
tzlocal==2.1
    # via streamlit
urllib3==1.26.6
    # via requests
validators==0.18.2
    # via streamlit
watchdog==2.1.3
    # via
    #   streamlit
    #   when-changed
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
when-changed==0.3.0
    # via -r requirements.in
widgetsnbextension==3.5.1
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
